move collate
