+++ docs/source/api_ref_data.rst
@@ -63,6 +63,17 @@ Converts data from common schema and conversation JSON formats into a list of to
     JSONToMessages
     ChosenRejectedToMessages
 
+Collaters
+---------
+
+Collaters used to collect samples into batches and handle any padding.
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    padded_collate
+
 Helper functions
 ----------------
