+++ recipes/full_finetune_single_device.py
@@ -433,7 +433,7 @@ class FullFinetuneRecipeSingleDevice(FTRecipeInterface):
             batch_size=batch_size,
             sampler=sampler,
             collate_fn=partial(
-                utils.padded_collate,
+                padded_collate,
                 padding_idx=self._tokenizer.pad_id,
                 ignore_idx=self._loss_fn.ignore_index,
             )
