+++ torchtune/utils/__init__.py
@@ -43,7 +43,6 @@ from ._profiler import (
 )
 from ._version import torch_version_ge
 from .argparse import TuneRecipeArgumentParser
-from .collate import padded_collate
 from .constants import (  # noqa
     ADAPTER_CONFIG,
     ADAPTER_KEY,
@@ -85,7 +84,6 @@ __all__ = [
     "is_distributed",
     "lora_fsdp_wrap_policy",
     "get_full_finetune_fsdp_wrap_policy",
-    "padded_collate",
     "get_unmasked_sequence_lengths",
     "set_activation_checkpointing",
     "set_default_dtype",
